test: extract agent machine suites
