+++ crates/agent-engine/src/agent_machine.rs
@@ -1286,99 +1286,6 @@ impl AgentMachine {
     }
 }
 
-/// Truncate a JSON payload to prevent context overflow
-/// Recursively truncates large string values while preserving structure
-#[cfg(test)]
-fn truncate_payload(payload: serde_json::Value, max_size: usize) -> serde_json::Value {
-    let payload_str = payload.to_string();
-    if payload_str.len() <= max_size {
-        return payload;
-    }
-
-    match payload {
-        serde_json::Value::Object(map) => {
-            let mut truncated = serde_json::Map::new();
-            let mut current_size = 0;
-
-            for (key, value) in map {
-                // Always include critical fields
-                let is_critical = matches!(key.as_str(), "success" | "error" | "url" | "title");
-
-                if is_critical {
-                    truncated.insert(key, value);
-                    continue;
-                }
-
-                // For content/aggregated_content fields, truncate aggressively
-                let processed_value = if key == "content" || key == "aggregated_content" {
-                    if let serde_json::Value::String(s) = &value {
-                        let truncated_str = truncate_text(s, max_size / 4);
-                        serde_json::Value::String(truncated_str)
-                    } else {
-                        value
-                    }
-                } else {
-                    truncate_payload(value, max_size / 2)
-                };
-
-                let value_str = processed_value.to_string();
-                if current_size + value_str.len() < max_size * 3 / 4 {
-                    truncated.insert(key, processed_value);
-                    current_size += value_str.len();
-                } else {
-                    truncated.insert(
-                        "_truncated".to_string(),
-                        serde_json::Value::String("Additional fields omitted".to_string()),
-                    );
-                    break;
-                }
-            }
-
-            serde_json::Value::Object(truncated)
-        }
-        serde_json::Value::Array(arr) => {
-            let arr_len = arr.len();
-            let mut truncated = Vec::new();
-            let mut current_size = 0;
-
-            for item in arr {
-                let processed = truncate_payload(item, max_size / arr_len.max(1));
-                let item_str = processed.to_string();
-
-                if current_size + item_str.len() < max_size * 3 / 4 {
-                    truncated.push(processed);
-                    current_size += item_str.len();
-                } else {
-                    truncated.push(serde_json::json!({
-                        "_note": "Additional array items omitted"
-                    }));
-                    break;
-                }
-            }
-
-            serde_json::Value::Array(truncated)
-        }
-        serde_json::Value::String(s) => {
-            if s.len() > max_size / 10 {
-                serde_json::Value::String(truncate_text(&s, max_size / 10))
-            } else {
-                serde_json::Value::String(s)
-            }
-        }
-        other => other,
-    }
-}
-
-/// Truncate text to a maximum length, adding ellipsis if truncated
-#[cfg(test)]
-fn truncate_text(text: &str, max_len: usize) -> String {
-    if text.len() <= max_len {
-        return text.to_string();
-    }
-    let truncated: String = text.chars().take(max_len).collect();
-    format!("{}...[truncated]", truncated)
-}
-
 fn fingerprint_turn_context_observation(
     model: &str,
     reasoning_effort: Option<alan_agent_protocol::ReasoningEffort>,
@@ -1444,2218 +1351,5 @@ pub fn latest_memory_flush_attempt_from_rollout_items(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::rollout::{
-        AgentMachineMeta, CheckpointRecord, CompactedItem, EffectRecord, EffectStatus, EventRecord,
-        MessageRecord, RolloutItem, RolloutRecorder,
-    };
-    use crate::tape::{ContentPart, ToolResponse};
-    use alan_agent_protocol::{
-        CompactionAttemptSnapshot, CompactionMode, CompactionReason, CompactionRequestMetadata,
-        CompactionResult, CompactionTrigger, MemoryFlushAttemptSnapshot, MemoryFlushResult,
-        MemoryFlushSkipReason,
-    };
-    use tempfile::TempDir;
-
-    #[test]
-    fn test_agent_machine_new() {
-        let machine = AgentMachine::new();
-        assert!(machine.tape.messages().is_empty());
-        assert!(machine.recorder.is_none());
-    }
-
-    #[test]
-    fn test_agent_machine_default() {
-        let machine = AgentMachine::default();
-        assert!(machine.tape.messages().is_empty());
-    }
-
-    #[test]
-    fn test_add_user_message() {
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("Hello, agent!");
-
-        let messages = machine.tape.messages();
-        assert_eq!(messages.len(), 1);
-        assert_eq!(messages[0].role(), MessageRole::User);
-        assert_eq!(messages[0].text_content(), "Hello, agent!");
-        assert_eq!(machine.user_turn_ordinal(), 1);
-    }
-
-    #[test]
-    fn test_user_turn_ordinal_is_monotonic_across_rollback() {
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("u1");
-        machine.add_user_message("u2");
-        assert_eq!(machine.user_turn_ordinal(), 2);
-
-        let removed = machine.rollback_last_turns(1);
-        assert!(removed.removed_messages > 0);
-        assert_eq!(removed.removed_turns, 1);
-        assert_eq!(machine.user_turn_count(), 1);
-        assert_eq!(machine.user_turn_ordinal(), 2);
-
-        machine.add_user_message("u3");
-        assert_eq!(machine.user_turn_count(), 2);
-        assert_eq!(machine.user_turn_ordinal(), 3);
-    }
-
-    #[test]
-    fn test_user_control_message_does_not_increment_turn_ordinal() {
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("u1");
-        assert_eq!(machine.user_turn_ordinal(), 1);
-
-        machine.add_user_control_message_parts(vec![ContentPart::structured(
-            serde_json::json!({"choice":"approve"}),
-        )]);
-
-        assert_eq!(machine.user_turn_count(), 2);
-        assert_eq!(machine.user_turn_ordinal(), 1);
-    }
-
-    #[test]
-    fn test_add_assistant_message() {
-        let mut machine = AgentMachine::new();
-        machine.add_assistant_message("I can help you!", None);
-
-        let messages = machine.tape.messages();
-        assert_eq!(messages.len(), 1);
-        assert_eq!(messages[0].role(), MessageRole::Assistant);
-        assert_eq!(messages[0].text_content(), "I can help you!");
-    }
-
-    #[test]
-    fn test_add_tool_message() {
-        let mut machine = AgentMachine::new();
-        let payload = serde_json::json!({"result": "success"});
-        machine.add_tool_message("call_123", "search_tool", payload);
-
-        let messages = machine.tape.messages();
-        assert_eq!(messages.len(), 1);
-        assert_eq!(messages[0].role(), MessageRole::Tool);
-        let responses = messages[0].tool_responses();
-        assert_eq!(responses.len(), 1);
-        assert_eq!(responses[0].id, "call_123");
-    }
-
-    #[test]
-    fn test_add_tool_message_accepts_content_parts_payload() {
-        let mut machine = AgentMachine::new();
-        let payload = serde_json::json!({
-            "content_parts": [
-                {"type": "text", "text": "hello"},
-                {"type": "attachment", "hash": "abc123", "mime_type": "image/png", "metadata": {"w": 10, "h": 10}}
-            ]
-        });
-        machine.add_tool_message("call_123", "capture", payload);
-
-        let messages = machine.tape.messages();
-        assert_eq!(messages.len(), 1);
-        let responses = messages[0].tool_responses();
-        assert_eq!(responses.len(), 1);
-        assert_eq!(responses[0].id, "call_123");
-        assert!(matches!(
-            responses[0].content.first(),
-            Some(ContentPart::Text { text }) if text == "hello"
-        ));
-        assert!(matches!(
-            responses[0].content.get(1),
-            Some(ContentPart::Attachment { hash, mime_type, .. })
-            if hash == "abc123" && mime_type == "image/png"
-        ));
-    }
-
-    #[test]
-    fn test_add_tool_message_accepts_content_parts_array_payload() {
-        let mut machine = AgentMachine::new();
-        let payload = serde_json::json!([
-            {"type": "text", "text": "part-a"},
-            {"type": "structured", "data": {"k": "v"}}
-        ]);
-        machine.add_tool_message("call_124", "custom", payload);
-
-        let messages = machine.tape.messages();
-        assert_eq!(messages.len(), 1);
-        let responses = messages[0].tool_responses();
-        assert_eq!(responses.len(), 1);
-        assert_eq!(responses[0].content.len(), 2);
-        assert!(matches!(
-            responses[0].content.first(),
-            Some(ContentPart::Text { text }) if text == "part-a"
-        ));
-        assert!(matches!(
-            responses[0].content.get(1),
-            Some(ContentPart::Structured { data }) if data["k"] == "v"
-        ));
-    }
-
-    #[test]
-    fn test_responses_continuation_can_be_marked_and_cleared() {
-        let mut machine = AgentMachine::new();
-        machine.mark_responses_continuation("openai_responses", "resp_123", 2, 7);
-
-        let continuation = machine.responses_continuation().expect("continuation");
-        assert_eq!(continuation.provider, "openai_responses");
-        assert_eq!(continuation.last_response_id, "resp_123");
-        assert_eq!(continuation.boundary_message_count, 2);
-        assert_eq!(continuation.reference_context_revision, 7);
-
-        machine.clear_responses_continuation("test");
-        assert!(machine.responses_continuation().is_none());
-    }
-
-    #[test]
-    fn test_multiple_messages() {
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("First");
-        machine.add_assistant_message("Second", None);
-        machine.add_user_message("Third");
-
-        let messages = machine.tape.messages();
-        assert_eq!(messages.len(), 3);
-        assert_eq!(messages[0].role(), MessageRole::User);
-        assert_eq!(messages[1].role(), MessageRole::Assistant);
-        assert_eq!(messages[2].role(), MessageRole::User);
-    }
-
-    #[test]
-    fn test_clear_machine() {
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("Test");
-
-        machine.clear();
-
-        assert!(machine.tape.messages().is_empty());
-    }
-
-    #[test]
-    fn test_message_role_serialization() {
-        let roles = vec![
-            (MessageRole::System, "\"system\""),
-            (MessageRole::Context, "\"context\""),
-            (MessageRole::User, "\"user\""),
-            (MessageRole::Assistant, "\"assistant\""),
-            (MessageRole::Tool, "\"tool\""),
-        ];
-
-        for (role, expected) in roles {
-            let json = serde_json::to_string(&role).unwrap();
-            assert_eq!(json, expected);
-
-            let deserialized: MessageRole = serde_json::from_str(expected).unwrap();
-            assert!(std::mem::discriminant(&deserialized) == std::mem::discriminant(&role));
-        }
-    }
-
-    #[test]
-    fn test_message_serialization() {
-        let message = Message::user("Hello");
-
-        let json = serde_json::to_string(&message).unwrap();
-        assert!(json.contains("Hello"));
-        assert!(json.contains("user"));
-
-        let deserialized: Message = serde_json::from_str(&json).unwrap();
-        assert_eq!(deserialized.text_content(), "Hello");
-    }
-
-    #[test]
-    fn test_message_serialization_with_tool() {
-        let message = Message::Tool {
-            responses: vec![ToolResponse {
-                id: "web_search".to_string(),
-                content: vec![ContentPart::structured(
-                    serde_json::json!({"result": "found"}),
-                )],
-            }],
-        };
-
-        let json = serde_json::to_string(&message).unwrap();
-        assert!(json.contains("web_search"));
-        assert!(json.contains("found"));
-
-        let deserialized: Message = serde_json::from_str(&json).unwrap();
-        assert_eq!(deserialized.tool_responses()[0].id, "web_search");
-    }
-
-    #[test]
-    fn test_machine_rollout_path_without_recorder() {
-        let machine = AgentMachine::new();
-        assert!(machine.rollout_path().is_none());
-    }
-
-    #[test]
-    fn test_machine_has_active_task_defaults_false() {
-        let machine = AgentMachine::new();
-        assert!(!machine.has_active_task);
-    }
-
-    #[test]
-    fn test_machine_clear_resets_active_task() {
-        let mut machine = AgentMachine::new();
-        machine.has_active_task = true;
-        machine.clear();
-        assert!(!machine.has_active_task);
-    }
-
-    #[test]
-    fn test_load_from_rollout_prefers_rich_message_payload_when_available() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-            let rollout_path = temp_dir.path().join("rollout-rich-message.jsonl");
-
-            let items = [
-                RolloutItem::AgentMachineMeta(AgentMachineMeta {
-                    rollout_id: "test-rich-rollout".to_string(),
-                    process_path: "/proc/test".to_string(),
-                    started_at: "2026-01-29T14:30:52Z".to_string(),
-                    cwd: "/tmp".to_string(),
-                    model: "gemini-2.0-flash".to_string(),
-                    reasoning_effort: None,
-                }),
-                RolloutItem::Message(MessageRecord {
-                    role: "assistant".to_string(),
-                    content: Some("final answer".to_string()),
-                    tool_name: None,
-                    message: Some(Message::Assistant {
-                        parts: vec![
-                            ContentPart::thinking("internal reasoning"),
-                            ContentPart::text("final answer"),
-                        ],
-                        tool_requests: vec![crate::tape::ToolRequest {
-                            id: "call_123".to_string(),
-                            name: "web_search".to_string(),
-                            arguments: serde_json::json!({"query":"alan"}),
-                        }],
-                    }),
-                    timestamp: "2026-01-29T14:30:56Z".to_string(),
-                }),
-            ];
-
-            let content = items
-                .iter()
-                .map(serde_json::to_string)
-                .collect::<Result<Vec<_>, _>>()
-                .unwrap()
-                .join("\n")
-                + "\n";
-            tokio::fs::write(&rollout_path, content).await.unwrap();
-            let machine = AgentMachine::load_from_rollout_in_dir(
-                &rollout_path,
-                "/proc/test",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-
-            assert_eq!(machine.tape.messages().len(), 1);
-            let message = &machine.tape.messages()[0];
-            assert_eq!(
-                message.thinking_content().as_deref(),
-                Some("internal reasoning")
-            );
-            assert_eq!(message.non_thinking_text_content(), "final answer");
-            assert_eq!(message.tool_requests().len(), 1);
-            assert_eq!(message.tool_requests()[0].name, "web_search");
-        });
-    }
-
-    #[test]
-    fn test_load_from_rollout_recovers_complete_records_before_torn_tail() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-            let rollout_path = temp_dir.path().join("rollout-torn-tail.jsonl");
-
-            let items = [
-                RolloutItem::AgentMachineMeta(AgentMachineMeta {
-                    rollout_id: "test-torn-tail".to_string(),
-                    process_path: "/proc/1".to_string(),
-                    started_at: "2026-01-29T14:30:52Z".to_string(),
-                    cwd: "/tmp".to_string(),
-                    model: "gemini-2.0-flash".to_string(),
-                    reasoning_effort: None,
-                }),
-                RolloutItem::Message(MessageRecord {
-                    role: "user".to_string(),
-                    content: Some("survives crash".to_string()),
-                    tool_name: None,
-                    message: Some(Message::user("survives crash")),
-                    timestamp: "2026-01-29T14:30:56Z".to_string(),
-                }),
-            ];
-            let mut content = items
-                .iter()
-                .map(serde_json::to_string)
-                .collect::<Result<Vec<_>, _>>()
-                .unwrap()
-                .join("\n");
-            content.push_str("\n{\"type\":\"message\",\"role\":");
-            tokio::fs::write(&rollout_path, content).await.unwrap();
-
-            let machine = AgentMachine::load_from_rollout_in_dir(
-                &rollout_path,
-                "/proc/2",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-
-            assert_eq!(machine.tape.messages(), &[Message::user("survives crash")]);
-        });
-    }
-
-    #[test]
-    fn test_load_from_rollout_does_not_count_runtime_confirmation_control_messages_as_turns() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-            let rollout_path = temp_dir.path().join("rollout-control-turn-ordinal.jsonl");
-
-            let items = [
-                RolloutItem::AgentMachineMeta(AgentMachineMeta {
-                    rollout_id: "test-control-turn-ordinal".to_string(),
-                    process_path: "/proc/test".to_string(),
-                    started_at: "2026-01-29T14:30:52Z".to_string(),
-                    cwd: "/tmp".to_string(),
-                    model: "gemini-2.0-flash".to_string(),
-                reasoning_effort: None,
-                }),
-                RolloutItem::Message(MessageRecord {
-                    role: "user".to_string(),
-                    content: Some("run task".to_string()),
-                    tool_name: None,
-                    message: Some(Message::User {
-                        parts: vec![ContentPart::text("run task")],
-                    }),
-                    timestamp: "2026-01-29T14:30:53Z".to_string(),
-                }),
-                RolloutItem::Message(MessageRecord {
-                    role: "user".to_string(),
-                    content: Some(
-                        "{\"checkpoint_id\":\"tool_escalation_call-1\",\"checkpoint_type\":\"tool_escalation\",\"choice\":\"approve\"}".to_string(),
-                    ),
-                    tool_name: None,
-                    message: Some(Message::User {
-                        parts: vec![ContentPart::structured(serde_json::json!({
-                            "checkpoint_id": "tool_escalation_call-1",
-                            "checkpoint_type": "tool_escalation",
-                            "choice": "approve",
-                            "__alan_internal_control": {
-                                "kind": "tool_escalation_confirmation",
-                                "version": 1,
-                                "source": "runtime/submission_handlers"
-                            }
-                        }))],
-                    }),
-                    timestamp: "2026-01-29T14:30:54Z".to_string(),
-                }),
-                RolloutItem::Checkpoint(CheckpointRecord {
-                    checkpoint_id: "tool_escalation_call-1".to_string(),
-                    checkpoint_type: "tool_escalation".to_string(),
-                    summary: "approve side effect".to_string(),
-                    choice: Some("approved".to_string()),
-                    knowledge_root: None,
-                    timestamp: "2026-01-29T14:30:54Z".to_string(),
-                }),
-                RolloutItem::Message(MessageRecord {
-                    role: "user".to_string(),
-                    content: Some("next task".to_string()),
-                    tool_name: None,
-                    message: Some(Message::user("next task")),
-                    timestamp: "2026-01-29T14:30:55Z".to_string(),
-                }),
-                RolloutItem::Message(MessageRecord {
-                    role: "user".to_string(),
-                    content: Some(
-                        "{\"checkpoint_id\":\"effect_replay_call-2\",\"checkpoint_type\":\"effect_replay_confirmation\",\"choice\":\"reject\",\"__alan_internal_control\":{\"kind\":\"effect_replay_confirmation\",\"version\":1,\"source\":\"runtime/submission_handlers\"}}"
-                            .to_string(),
-                    ),
-                    tool_name: None,
-                    message: Some(Message::user_parts(vec![ContentPart::structured(
-                        serde_json::json!({
-                            "checkpoint_id": "effect_replay_call-2",
-                            "checkpoint_type": "effect_replay_confirmation",
-                            "choice": "reject",
-                            "__alan_internal_control": {
-                                "kind": "effect_replay_confirmation",
-                                "version": 1,
-                                "source": "runtime/submission_handlers"
-                            }
-                        }),
-                    )])),
-                    timestamp: "2026-01-29T14:30:56Z".to_string(),
-                }),
-                RolloutItem::Checkpoint(CheckpointRecord {
-                    checkpoint_id: "effect_replay_call-2".to_string(),
-                    checkpoint_type: "effect_replay_confirmation".to_string(),
-                    summary: "reject side effect".to_string(),
-                    choice: Some("rejected".to_string()),
-                    knowledge_root: None,
-                    timestamp: "2026-01-29T14:30:56Z".to_string(),
-                }),
-            ];
-
-            let content = items
-                .iter()
-                .map(serde_json::to_string)
-                .collect::<Result<Vec<_>, _>>()
-                .unwrap()
-                .join("\n")
-                + "\n";
-
-            tokio::fs::write(&rollout_path, content).await.unwrap();
-            let mut machine = AgentMachine::load_from_rollout_in_dir(
-                &rollout_path,
-                "/proc/test",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-
-            assert_eq!(
-                machine.user_turn_ordinal(),
-                2,
-                "only non-control user messages should increment turn ordinal during recovery"
-            );
-            assert_eq!(machine.user_turn_count(), 4);
-            let removed = machine.rollback_last_turns(2);
-            assert_eq!(
-                removed.removed_messages, 4,
-                "runtime control messages should remain outside logical user-turn rollback"
-            );
-            assert_eq!(removed.removed_turns, 2);
-        });
-    }
-
-    #[test]
-    fn test_load_from_rollout_excludes_current_runtime_control_from_turn_ordinal() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-            let rollout_path = temp_dir
-                .path()
-                .join("rollout-strict-control-without-checkpoint.jsonl");
-
-            let items = [
-                RolloutItem::AgentMachineMeta(AgentMachineMeta {
-                    rollout_id: "test-strict-control-without-checkpoint".to_string(),
-                    process_path: "/proc/test".to_string(),
-                    started_at: "2026-01-29T14:30:52Z".to_string(),
-                    cwd: "/tmp".to_string(),
-                    model: "gemini-2.0-flash".to_string(),
-                    reasoning_effort: None,
-                }),
-                RolloutItem::Message(MessageRecord {
-                    role: "user".to_string(),
-                    content: Some(
-                        "{\"checkpoint_id\":\"tool_escalation_call-11\",\"checkpoint_type\":\"tool_escalation\",\"choice\":\"approve\",\"__alan_internal_control\":{\"kind\":\"tool_escalation_confirmation\",\"version\":1,\"source\":\"runtime/submission_handlers\"}}"
-                            .to_string(),
-                    ),
-                    tool_name: None,
-                    message: Some(Message::User {
-                        parts: vec![ContentPart::structured(serde_json::json!({
-                            "checkpoint_id": "tool_escalation_call-11",
-                            "checkpoint_type": "tool_escalation",
-                            "choice": "approve",
-                            "__alan_internal_control": {
-                                "kind": "tool_escalation_confirmation",
-                                "version": 1,
-                                "source": "runtime/submission_handlers"
-                            }
-                        }))],
-                    }),
-                    timestamp: "2026-01-29T14:30:53Z".to_string(),
-                }),
-            ];
-
-            let content = items
-                .iter()
-                .map(serde_json::to_string)
-                .collect::<Result<Vec<_>, _>>()
-                .unwrap()
-                .join("\n")
-                + "\n";
-
-            tokio::fs::write(&rollout_path, content).await.unwrap();
-            let machine = AgentMachine::load_from_rollout_in_dir(
-                &rollout_path,
-                "/proc/test",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-
-            assert_eq!(machine.user_turn_ordinal(), 0);
-            assert_eq!(machine.user_turn_count(), 1);
-        });
-    }
-
-    #[test]
-    fn test_load_from_rollout_counts_user_payloads_without_internal_control_marker() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-            let rollout_path = temp_dir.path().join("rollout-user-payload-turn-ordinal.jsonl");
-
-            let items = [
-                RolloutItem::AgentMachineMeta(AgentMachineMeta {
-                    rollout_id: "test-user-payload-turn-ordinal".to_string(),
-                    process_path: "/proc/test".to_string(),
-                    started_at: "2026-01-29T14:30:52Z".to_string(),
-                    cwd: "/tmp".to_string(),
-                    model: "gemini-2.0-flash".to_string(),
-                    reasoning_effort: None,
-                }),
-                RolloutItem::Message(MessageRecord {
-                    role: "user".to_string(),
-                    content: Some(
-                        "{\"checkpoint_id\":\"custom-id\",\"checkpoint_type\":\"tool_escalation\",\"choice\":\"approve\"}"
-                            .to_string(),
-                    ),
-                    tool_name: None,
-                    message: Some(Message::User {
-                        parts: vec![ContentPart::structured(serde_json::json!({
-                            "checkpoint_id": "custom-id",
-                            "checkpoint_type": "tool_escalation",
-                            "choice": "approve",
-                        }))],
-                    }),
-                    timestamp: "2026-01-29T14:30:53Z".to_string(),
-                }),
-                RolloutItem::Message(MessageRecord {
-                    role: "user".to_string(),
-                    content: Some(
-                        "{\"checkpoint_id\":\"manual-id\",\"checkpoint_type\":\"tool_escalation\",\"choice\":\"reject\"}"
-                            .to_string(),
-                    ),
-                    tool_name: None,
-                    message: Some(Message::user_parts(vec![ContentPart::structured(
-                        serde_json::json!({
-                            "checkpoint_id": "manual-id",
-                            "checkpoint_type": "tool_escalation",
-                            "choice": "reject"
-                        }),
-                    )])),
-                    timestamp: "2026-01-29T14:30:54Z".to_string(),
-                }),
-            ];
-
-            let content = items
-                .iter()
-                .map(serde_json::to_string)
-                .collect::<Result<Vec<_>, _>>()
-                .unwrap()
-                .join("\n")
-                + "\n";
-
-            tokio::fs::write(&rollout_path, content).await.unwrap();
-            let machine = AgentMachine::load_from_rollout_in_dir(
-                &rollout_path,
-                "/proc/test",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-
-            assert_eq!(
-                machine.user_turn_ordinal(),
-                2,
-                "user payloads without internal control markers should count as turns"
-            );
-            assert_eq!(machine.user_turn_count(), 2);
-        });
-    }
-
-    #[test]
-    fn test_load_from_rollout_preserves_turn_ordinal_across_repeated_recovery() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-            let rollout_path = temp_dir.path().join("rollout-turn-ordinal-recovery.jsonl");
-
-            let items = [
-                RolloutItem::AgentMachineMeta(AgentMachineMeta {
-                    rollout_id: "test-turn-ordinal-recovery".to_string(),
-                    process_path: "/proc/test".to_string(),
-                    started_at: "2026-01-29T14:30:52Z".to_string(),
-                    cwd: "/tmp".to_string(),
-                    model: "gemini-2.0-flash".to_string(),
-                    reasoning_effort: None,
-                }),
-                RolloutItem::Message(MessageRecord {
-                    role: "user".to_string(),
-                    content: Some("task one".to_string()),
-                    tool_name: None,
-                    message: Some(Message::user("task one")),
-                    timestamp: "2026-01-29T14:30:53Z".to_string(),
-                }),
-                RolloutItem::Message(MessageRecord {
-                    role: "assistant".to_string(),
-                    content: Some("ack".to_string()),
-                    tool_name: None,
-                    message: Some(Message::assistant("ack")),
-                    timestamp: "2026-01-29T14:30:54Z".to_string(),
-                }),
-                RolloutItem::Message(MessageRecord {
-                    role: "user".to_string(),
-                    content: Some("task two".to_string()),
-                    tool_name: None,
-                    message: Some(Message::user("task two")),
-                    timestamp: "2026-01-29T14:30:55Z".to_string(),
-                }),
-            ];
-
-            let content = items
-                .iter()
-                .map(serde_json::to_string)
-                .collect::<Result<Vec<_>, _>>()
-                .unwrap()
-                .join("\n")
-                + "\n";
-
-            tokio::fs::write(&rollout_path, content).await.unwrap();
-
-            let first = AgentMachine::load_from_rollout_in_dir(
-                &rollout_path,
-                "/proc/test",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-            assert_eq!(first.user_turn_ordinal(), 2);
-            drop(first);
-
-            let second = AgentMachine::load_from_rollout_in_dir(
-                &rollout_path,
-                "/proc/test",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-            assert_eq!(
-                second.user_turn_ordinal(),
-                2,
-                "recovered history should preserve monotonic turn ordinal across repeated recovery"
-            );
-        });
-    }
-
-    #[test]
-    fn test_load_from_rollout_preserves_turn_ordinal_floor_from_effect_keys_after_compaction() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-            let rollout_path = temp_dir.path().join("rollout-compaction-turn-floor.jsonl");
-
-            let items = [
-                RolloutItem::AgentMachineMeta(AgentMachineMeta {
-                    rollout_id: "sess-compaction-floor".to_string(),
-                    process_path: "/proc/test".to_string(),
-                    started_at: "2026-01-29T14:30:52Z".to_string(),
-                    cwd: "/tmp".to_string(),
-                    model: "gemini-2.0-flash".to_string(),
-                    reasoning_effort: None,
-                }),
-                RolloutItem::Compacted(CompactedItem {
-                    message: "Older turns compacted".to_string(),
-                    attempt_id: None,
-                    trigger: None,
-                    reason: None,
-                    focus: None,
-                    input_messages: None,
-                    output_messages: None,
-                    input_tokens: None,
-                    output_tokens: None,
-                    duration_ms: None,
-                    retry_count: None,
-                    result: None,
-                    reference_context_revision: None,
-                    timestamp: "2026-01-29T14:31:00Z".to_string(),
-                }),
-                RolloutItem::Message(MessageRecord {
-                    role: "user".to_string(),
-                    content: Some("latest visible turn".to_string()),
-                    tool_name: None,
-                    message: Some(Message::user("latest visible turn")),
-                    timestamp: "2026-01-29T14:31:01Z".to_string(),
-                }),
-                RolloutItem::Effect(EffectRecord {
-                    effect_id: "ef-compaction".to_string(),
-                    process_path: "sess-compaction-floor".to_string(),
-                    tool_call_id: "call-1".to_string(),
-                    idempotency_key: "machine:turn:7:fp-1".to_string(),
-                    effect_type: "file".to_string(),
-                    request_fingerprint: "fp-1".to_string(),
-                    result_digest: Some("digest-1".to_string()),
-                    result_payload: Some(serde_json::json!({"ok": true})),
-                    status: EffectStatus::Applied,
-                    applied_at: Some("2026-01-29T14:31:02Z".to_string()),
-                    reason: None,
-                    dedupe_hit: false,
-                    timestamp: "2026-01-29T14:31:02Z".to_string(),
-                }),
-            ];
-
-            let content = items
-                .iter()
-                .map(serde_json::to_string)
-                .collect::<Result<Vec<_>, _>>()
-                .unwrap()
-                .join("\n")
-                + "\n";
-            tokio::fs::write(&rollout_path, content).await.unwrap();
-
-            let machine = AgentMachine::load_from_rollout_in_dir(
-                &rollout_path,
-                "/proc/test",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-
-            assert_eq!(
-                machine.user_turn_ordinal(),
-                7,
-                "effect idempotency keys should preserve turn ordinal floor after compaction"
-            );
-            assert_eq!(machine.user_turn_count(), 1);
-        });
-    }
-
-    #[test]
-    fn test_load_from_rollout_preserves_generic_event_records_across_recovery() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-            let rollout_path = temp_dir.path().join("rollout-events.jsonl");
-
-            let items = [
-                RolloutItem::AgentMachineMeta(AgentMachineMeta {
-                    rollout_id: "sess-events".to_string(),
-                    process_path: "/proc/test".to_string(),
-                    started_at: "2026-01-29T14:30:52Z".to_string(),
-                    cwd: "/tmp".to_string(),
-                    model: "gemini-2.0-flash".to_string(),
-                    reasoning_effort: None,
-                }),
-                RolloutItem::Event(EventRecord {
-                    event_type: "custom_event".to_string(),
-                    payload: serde_json::json!({
-                        "phase": "testing",
-                        "value": 5
-                    }),
-                    timestamp: "2026-01-29T14:31:00Z".to_string(),
-                }),
-            ];
-
-            let content = items
-                .iter()
-                .map(serde_json::to_string)
-                .collect::<Result<Vec<_>, _>>()
-                .unwrap()
-                .join("\n")
-                + "\n";
-            tokio::fs::write(&rollout_path, content).await.unwrap();
-
-            let machine = AgentMachine::load_from_rollout_in_dir(
-                &rollout_path,
-                "/proc/test",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-            machine.flush().await;
-
-            let recovered_path = machine
-                .rollout_path()
-                .expect("recovered machine should have rollout path")
-                .clone();
-            let recovered_items = RolloutRecorder::load_history(&recovered_path)
-                .await
-                .unwrap();
-
-            let event = recovered_items.into_iter().find_map(|item| match item {
-                RolloutItem::Event(event) if event.event_type == "custom_event" => Some(event),
-                _ => None,
-            });
-
-            let event = event.expect("expected recovered custom event");
-            assert_eq!(event.payload["phase"], "testing");
-            assert_eq!(event.payload["value"], 5);
-            assert_eq!(event.timestamp, "2026-01-29T14:31:00Z");
-        });
-    }
-
-    #[test]
-    fn test_load_from_rollout_repersists_memory_flush_attempt_records() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-            let rollout_path = temp_dir.path().join("rollout-memory-flush-attempt.jsonl");
-
-            let attempt = MemoryFlushAttemptSnapshot {
-                attempt_id: "flush-123".to_string(),
-                compaction_mode: CompactionMode::AutoPreTurn,
-                pressure_level: alan_agent_protocol::CompactionPressureLevel::Soft,
-                result: MemoryFlushResult::Success,
-                skip_reason: None,
-                source_messages: Some(7),
-                output_path: Some(".alan/memory/daily/2026-03-03.md".to_string()),
-                warning_message: None,
-                error_message: None,
-                timestamp: "2026-03-03T10:00:00Z".to_string(),
-            };
-
-            let items = [
-                RolloutItem::AgentMachineMeta(AgentMachineMeta {
-                    rollout_id: "sess-memory-flush-attempt".to_string(),
-                    process_path: "/proc/test".to_string(),
-                    started_at: "2026-03-03T09:59:52Z".to_string(),
-                    cwd: "/tmp".to_string(),
-                    model: "gemini-2.0-flash".to_string(),
-                    reasoning_effort: None,
-                }),
-                RolloutItem::MemoryFlushAttempt(attempt.clone()),
-            ];
-
-            let content = items
-                .iter()
-                .map(serde_json::to_string)
-                .collect::<Result<Vec<_>, _>>()
-                .unwrap()
-                .join("\n")
-                + "\n";
-            tokio::fs::write(&rollout_path, content).await.unwrap();
-
-            let machine = AgentMachine::load_from_rollout_in_dir(
-                &rollout_path,
-                "/proc/test",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-
-            assert_eq!(machine.latest_memory_flush_attempt(), Some(&attempt));
-
-            machine.flush().await;
-            let recovered_path = machine
-                .rollout_path()
-                .expect("recovered machine should have rollout path")
-                .clone();
-            let recovered_items = RolloutRecorder::load_history(&recovered_path)
-                .await
-                .unwrap();
-
-            let persisted = recovered_items.into_iter().find_map(|item| match item {
-                RolloutItem::MemoryFlushAttempt(snapshot) => Some(snapshot),
-                _ => None,
-            });
-
-            assert_eq!(persisted, Some(attempt));
-        });
-    }
-
-    #[test]
-    fn test_load_from_rollout_restores_latest_compaction_attempt_item_when_summary_is_persisted() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-            let rollout_path = temp_dir.path().join("rollout-compaction-attempt.jsonl");
-
-            let attempt = CompactionAttemptSnapshot {
-                attempt_id: "attempt-123".to_string(),
-                submission_id: None,
-                request: CompactionRequestMetadata {
-                    mode: CompactionMode::AutoPreTurn,
-                    trigger: CompactionTrigger::Auto,
-                    reason: CompactionReason::WindowPressure,
-                    focus: None,
-                },
-                result: CompactionResult::Retry,
-                pressure_level: None,
-                memory_flush_attempt_id: None,
-                input_messages: Some(18),
-                output_messages: Some(5),
-                input_prompt_tokens: Some(1500),
-                output_prompt_tokens: Some(480),
-                retry_count: 1,
-                tape_mutated: true,
-                warning_message: None,
-                error_message: None,
-                failure_streak: None,
-                reference_context_revision_before: Some(4),
-                reference_context_revision_after: Some(4),
-                timestamp: "2026-01-29T14:31:00Z".to_string(),
-            };
-
-            let items = [
-                RolloutItem::AgentMachineMeta(AgentMachineMeta {
-                    rollout_id: "sess-compaction-attempt".to_string(),
-                    process_path: "/proc/test".to_string(),
-                    started_at: "2026-01-29T14:30:52Z".to_string(),
-                    cwd: "/tmp".to_string(),
-                    model: "gemini-2.0-flash".to_string(),
-                    reasoning_effort: None,
-                }),
-                RolloutItem::CompactionAttempt(attempt.clone()),
-                RolloutItem::Compacted(CompactedItem {
-                    message: "Summary after retry".to_string(),
-                    attempt_id: Some(attempt.attempt_id.clone()),
-                    trigger: Some(CompactionTrigger::Auto),
-                    reason: Some(CompactionReason::WindowPressure),
-                    focus: None,
-                    input_messages: Some(18),
-                    output_messages: Some(5),
-                    input_tokens: Some(1500),
-                    output_tokens: Some(480),
-                    duration_ms: Some(42),
-                    retry_count: Some(1),
-                    result: Some(CompactionResult::Retry),
-                    reference_context_revision: Some(4),
-                    timestamp: "2026-01-29T14:31:01Z".to_string(),
-                }),
-            ];
-
-            let content = items
-                .iter()
-                .map(serde_json::to_string)
-                .collect::<Result<Vec<_>, _>>()
-                .unwrap()
-                .join("\n")
-                + "\n";
-            tokio::fs::write(&rollout_path, content).await.unwrap();
-
-            let machine = AgentMachine::load_from_rollout_in_dir(
-                &rollout_path,
-                "/proc/test",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-
-            assert_eq!(machine.latest_compaction_attempt(), Some(&attempt));
-        });
-    }
-
-    #[test]
-    fn test_latest_compaction_attempt_from_rollout_matches_summary_by_attempt_id() {
-        let completed_attempt = CompactionAttemptSnapshot {
-            attempt_id: "attempt-complete".to_string(),
-            submission_id: None,
-            request: CompactionRequestMetadata {
-                mode: CompactionMode::Manual,
-                trigger: CompactionTrigger::Manual,
-                reason: CompactionReason::ExplicitRequest,
-                focus: Some("preserve tasks".to_string()),
-            },
-            result: CompactionResult::Success,
-            pressure_level: None,
-            memory_flush_attempt_id: None,
-            input_messages: Some(18),
-            output_messages: Some(5),
-            input_prompt_tokens: Some(1500),
-            output_prompt_tokens: Some(480),
-            retry_count: 0,
-            tape_mutated: true,
-            warning_message: None,
-            error_message: None,
-            failure_streak: None,
-            reference_context_revision_before: Some(4),
-            reference_context_revision_after: Some(5),
-            timestamp: "2026-01-29T14:31:00Z".to_string(),
-        };
-        let incomplete_retry = CompactionAttemptSnapshot {
-            attempt_id: "attempt-retry".to_string(),
-            submission_id: None,
-            request: CompactionRequestMetadata {
-                mode: CompactionMode::AutoPreTurn,
-                trigger: CompactionTrigger::Auto,
-                reason: CompactionReason::WindowPressure,
-                focus: None,
-            },
-            result: CompactionResult::Retry,
-            pressure_level: None,
-            memory_flush_attempt_id: None,
-            input_messages: Some(24),
-            output_messages: Some(8),
-            input_prompt_tokens: Some(1800),
-            output_prompt_tokens: Some(500),
-            retry_count: 1,
-            tape_mutated: true,
-            warning_message: None,
-            error_message: None,
-            failure_streak: None,
-            reference_context_revision_before: Some(5),
-            reference_context_revision_after: Some(5),
-            timestamp: "2026-01-29T14:32:00Z".to_string(),
-        };
-        let items = [
-            RolloutItem::CompactionAttempt(completed_attempt.clone()),
-            RolloutItem::CompactionAttempt(incomplete_retry),
-            RolloutItem::Compacted(CompactedItem {
-                message: "Summary after retry".to_string(),
-                attempt_id: Some(completed_attempt.attempt_id.clone()),
-                trigger: Some(CompactionTrigger::Manual),
-                reason: Some(CompactionReason::ExplicitRequest),
-                focus: Some("preserve tasks".to_string()),
-                input_messages: Some(18),
-                output_messages: Some(5),
-                input_tokens: Some(1500),
-                output_tokens: Some(480),
-                duration_ms: Some(42),
-                retry_count: Some(0),
-                result: Some(CompactionResult::Success),
-                reference_context_revision: Some(4),
-                timestamp: "2026-01-29T14:31:01Z".to_string(),
-            }),
-        ];
-
-        assert_eq!(
-            latest_compaction_attempt_from_rollout_items(&items),
-            Some(completed_attempt)
-        );
-    }
-
-    #[test]
-    fn test_latest_compaction_attempt_from_rollout_ignores_incomplete_tape_mutation() {
-        let failure = CompactionAttemptSnapshot {
-            attempt_id: "attempt-failure".to_string(),
-            submission_id: None,
-            request: CompactionRequestMetadata {
-                mode: CompactionMode::Manual,
-                trigger: CompactionTrigger::Manual,
-                reason: CompactionReason::ExplicitRequest,
-                focus: None,
-            },
-            result: CompactionResult::Failure,
-            pressure_level: None,
-            memory_flush_attempt_id: None,
-            input_messages: Some(18),
-            output_messages: None,
-            input_prompt_tokens: Some(1400),
-            output_prompt_tokens: None,
-            retry_count: 0,
-            tape_mutated: false,
-            warning_message: Some("Preserving existing context".to_string()),
-            error_message: Some("synthetic failure".to_string()),
-            failure_streak: Some(1),
-            reference_context_revision_before: Some(4),
-            reference_context_revision_after: None,
-            timestamp: "2026-01-29T14:31:00Z".to_string(),
-        };
-        let incomplete_retry = CompactionAttemptSnapshot {
-            attempt_id: "attempt-retry".to_string(),
-            submission_id: None,
-            request: CompactionRequestMetadata {
-                mode: CompactionMode::AutoPreTurn,
-                trigger: CompactionTrigger::Auto,
-                reason: CompactionReason::WindowPressure,
-                focus: None,
-            },
-            result: CompactionResult::Retry,
-            pressure_level: None,
-            memory_flush_attempt_id: None,
-            input_messages: Some(24),
-            output_messages: Some(8),
-            input_prompt_tokens: Some(1800),
-            output_prompt_tokens: Some(500),
-            retry_count: 1,
-            tape_mutated: true,
-            warning_message: None,
-            error_message: None,
-            failure_streak: None,
-            reference_context_revision_before: Some(5),
-            reference_context_revision_after: Some(5),
-            timestamp: "2026-01-29T14:32:00Z".to_string(),
-        };
-        let items = [
-            RolloutItem::CompactionAttempt(failure.clone()),
-            RolloutItem::CompactionAttempt(incomplete_retry),
-        ];
-
-        assert_eq!(
-            latest_compaction_attempt_from_rollout_items(&items),
-            Some(failure)
-        );
-    }
-
-    #[test]
-    fn test_latest_compaction_attempt_from_rollout_does_not_let_linked_summary_override_newer_attempt()
-     {
-        let completed_attempt = CompactionAttemptSnapshot {
-            attempt_id: "attempt-complete".to_string(),
-            submission_id: None,
-            request: CompactionRequestMetadata {
-                mode: CompactionMode::Manual,
-                trigger: CompactionTrigger::Manual,
-                reason: CompactionReason::ExplicitRequest,
-                focus: Some("preserve tasks".to_string()),
-            },
-            result: CompactionResult::Success,
-            pressure_level: None,
-            memory_flush_attempt_id: None,
-            input_messages: Some(18),
-            output_messages: Some(5),
-            input_prompt_tokens: Some(1500),
-            output_prompt_tokens: Some(480),
-            retry_count: 0,
-            tape_mutated: true,
-            warning_message: None,
-            error_message: None,
-            failure_streak: None,
-            reference_context_revision_before: Some(4),
-            reference_context_revision_after: Some(5),
-            timestamp: "2026-01-29T14:31:00Z".to_string(),
-        };
-        let failure = CompactionAttemptSnapshot {
-            attempt_id: "attempt-failure".to_string(),
-            submission_id: None,
-            request: CompactionRequestMetadata {
-                mode: CompactionMode::Manual,
-                trigger: CompactionTrigger::Manual,
-                reason: CompactionReason::ExplicitRequest,
-                focus: None,
-            },
-            result: CompactionResult::Failure,
-            pressure_level: None,
-            memory_flush_attempt_id: None,
-            input_messages: Some(18),
-            output_messages: None,
-            input_prompt_tokens: Some(1400),
-            output_prompt_tokens: None,
-            retry_count: 1,
-            tape_mutated: false,
-            warning_message: Some("Preserving existing context".to_string()),
-            error_message: Some("synthetic failure".to_string()),
-            failure_streak: Some(1),
-            reference_context_revision_before: Some(5),
-            reference_context_revision_after: None,
-            timestamp: "2026-01-29T14:32:00Z".to_string(),
-        };
-        let items = [
-            RolloutItem::CompactionAttempt(completed_attempt),
-            RolloutItem::CompactionAttempt(failure.clone()),
-            RolloutItem::Compacted(CompactedItem {
-                message: "Summary after retry".to_string(),
-                attempt_id: Some("attempt-complete".to_string()),
-                trigger: Some(CompactionTrigger::Manual),
-                reason: Some(CompactionReason::ExplicitRequest),
-                focus: Some("preserve tasks".to_string()),
-                input_messages: Some(18),
-                output_messages: Some(5),
-                input_tokens: Some(1500),
-                output_tokens: Some(480),
-                duration_ms: Some(42),
-                retry_count: Some(0),
-                result: Some(CompactionResult::Success),
-                reference_context_revision: Some(4),
-                timestamp: "2026-01-29T14:31:01Z".to_string(),
-            }),
-        ];
-
-        assert_eq!(
-            latest_compaction_attempt_from_rollout_items(&items),
-            Some(failure)
-        );
-    }
-
-    #[test]
-    fn test_load_from_rollout_does_not_let_repersisted_linked_summary_override_newer_failure() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-            let rollout_path = temp_dir.path().join("rollout-linked-summary.jsonl");
-
-            let completed_attempt = CompactionAttemptSnapshot {
-                attempt_id: "attempt-complete".to_string(),
-                submission_id: None,
-                request: CompactionRequestMetadata {
-                    mode: CompactionMode::Manual,
-                    trigger: CompactionTrigger::Manual,
-                    reason: CompactionReason::ExplicitRequest,
-                    focus: Some("preserve tasks".to_string()),
-                },
-                result: CompactionResult::Success,
-                pressure_level: None,
-                memory_flush_attempt_id: None,
-                input_messages: Some(18),
-                output_messages: Some(5),
-                input_prompt_tokens: Some(1500),
-                output_prompt_tokens: Some(480),
-                retry_count: 0,
-                tape_mutated: true,
-                warning_message: None,
-                error_message: None,
-                failure_streak: None,
-                reference_context_revision_before: Some(4),
-                reference_context_revision_after: Some(5),
-                timestamp: "2026-01-29T14:31:00Z".to_string(),
-            };
-            let failure = CompactionAttemptSnapshot {
-                attempt_id: "attempt-failure".to_string(),
-                submission_id: None,
-                request: CompactionRequestMetadata {
-                    mode: CompactionMode::Manual,
-                    trigger: CompactionTrigger::Manual,
-                    reason: CompactionReason::ExplicitRequest,
-                    focus: None,
-                },
-                result: CompactionResult::Failure,
-                pressure_level: None,
-                memory_flush_attempt_id: None,
-                input_messages: Some(18),
-                output_messages: None,
-                input_prompt_tokens: Some(1400),
-                output_prompt_tokens: None,
-                retry_count: 1,
-                tape_mutated: false,
-                warning_message: Some("Preserving existing context".to_string()),
-                error_message: Some("synthetic failure".to_string()),
-                failure_streak: Some(1),
-                reference_context_revision_before: Some(5),
-                reference_context_revision_after: None,
-                timestamp: "2026-01-29T14:32:00Z".to_string(),
-            };
-            let items = [
-                RolloutItem::AgentMachineMeta(AgentMachineMeta {
-                    rollout_id: "sess-linked-summary".to_string(),
-                    process_path: "/proc/test".to_string(),
-                    started_at: "2026-01-29T14:30:52Z".to_string(),
-                    cwd: "/tmp".to_string(),
-                    model: "gemini-2.0-flash".to_string(),
-                    reasoning_effort: None,
-                }),
-                RolloutItem::CompactionAttempt(completed_attempt.clone()),
-                RolloutItem::Compacted(CompactedItem {
-                    message: "Summary after retry".to_string(),
-                    attempt_id: Some(completed_attempt.attempt_id.clone()),
-                    trigger: Some(CompactionTrigger::Manual),
-                    reason: Some(CompactionReason::ExplicitRequest),
-                    focus: Some("preserve tasks".to_string()),
-                    input_messages: Some(18),
-                    output_messages: Some(5),
-                    input_tokens: Some(1500),
-                    output_tokens: Some(480),
-                    duration_ms: Some(42),
-                    retry_count: Some(0),
-                    result: Some(CompactionResult::Success),
-                    reference_context_revision: Some(4),
-                    timestamp: "2026-01-29T14:31:01Z".to_string(),
-                }),
-                RolloutItem::CompactionAttempt(failure.clone()),
-            ];
-
-            let content = items
-                .iter()
-                .map(serde_json::to_string)
-                .collect::<Result<Vec<_>, _>>()
-                .unwrap()
-                .join("\n")
-                + "\n";
-            tokio::fs::write(&rollout_path, content).await.unwrap();
-
-            let machine = AgentMachine::load_from_rollout_in_dir(
-                &rollout_path,
-                "/proc/test",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-
-            assert_eq!(machine.latest_compaction_attempt(), Some(&failure));
-
-            machine.flush().await;
-            let recovered_path = machine
-                .rollout_path()
-                .expect("recovered machine should have rollout path")
-                .clone();
-            let reloaded = AgentMachine::load_from_rollout_in_dir(
-                &recovered_path,
-                "/proc/test",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-
-            assert_eq!(reloaded.latest_compaction_attempt(), Some(&failure));
-        });
-    }
-
-    #[test]
-    fn test_empty_message_content() {
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("");
-
-        let messages = machine.tape.messages();
-        assert_eq!(messages[0].text_content(), "");
-    }
-
-    #[test]
-    fn test_unicode_message_content() {
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("你好，世界！🌍");
-
-        let messages = machine.tape.messages();
-        assert_eq!(messages[0].text_content(), "你好，世界！🌍");
-    }
-
-    #[test]
-    fn test_record_tool_call() {
-        let machine = AgentMachine::new();
-        let args = serde_json::json!({"query": "test"});
-        let result = serde_json::json!({"status": "ok"});
-
-        // Should not panic without recorder
-        machine.record_tool_call("search_tool", args, result, true);
-    }
-
-    #[test]
-    fn test_record_effect_updates_lookup_index() {
-        let mut machine = AgentMachine::new();
-        let effect = EffectRecord {
-            effect_id: "ef-1".to_string(),
-            process_path: "/proc/test".to_string(),
-            tool_call_id: "call-1".to_string(),
-            idempotency_key: "idem-1".to_string(),
-            effect_type: "file".to_string(),
-            request_fingerprint: "fp-1".to_string(),
-            result_digest: None,
-            result_payload: None,
-            status: EffectStatus::Unknown,
-            applied_at: None,
-            reason: Some("pending".to_string()),
-            dedupe_hit: false,
-            timestamp: "2026-03-03T10:00:00Z".to_string(),
-        };
-
-        machine.record_effect(effect);
-
-        let restored = machine.effect_by_idempotency_key("idem-1").unwrap();
-        assert_eq!(restored.effect_id, "ef-1");
-        assert_eq!(restored.status, EffectStatus::Unknown);
-    }
-
-    #[test]
-    fn test_record_checkpoint() {
-        let machine = AgentMachine::new();
-
-        // Should not panic without recorder
-        machine.record_checkpoint(
-            "cp-123",
-            "supplier_list",
-            "Test checkpoint",
-            Some("approve"),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_flush() {
-        let machine = AgentMachine::new();
-        // Should not panic without recorder
-        machine.flush().await;
-    }
-
-    #[test]
-    fn test_add_user_message_with_tool_name() {
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("Hello");
-        let messages = machine.tape.messages();
-        assert!(messages[0].is_user());
-    }
-
-    #[test]
-    fn test_record_event() {
-        let machine = AgentMachine::new();
-        // Should not panic without recorder
-        machine.record_event("test_event", serde_json::json!({"key": "value"}));
-    }
-
-    #[test]
-    fn test_record_summary() {
-        let machine = AgentMachine::new();
-        // Should not panic without recorder
-        machine.record_summary("Test summary");
-    }
-
-    #[test]
-    fn test_latest_memory_flush_attempt_from_rollout_items_returns_latest_attempt() {
-        let first = MemoryFlushAttemptSnapshot {
-            attempt_id: "flush-1".to_string(),
-            compaction_mode: CompactionMode::AutoPreTurn,
-            pressure_level: alan_agent_protocol::CompactionPressureLevel::Soft,
-            result: MemoryFlushResult::Skipped,
-            skip_reason: Some(MemoryFlushSkipReason::ReadOnlyMemoryDir),
-            source_messages: Some(4),
-            output_path: None,
-            warning_message: Some("memory dir is read-only".to_string()),
-            error_message: None,
-            timestamp: "2026-03-03T10:00:00Z".to_string(),
-        };
-        let second = MemoryFlushAttemptSnapshot {
-            attempt_id: "flush-2".to_string(),
-            compaction_mode: CompactionMode::AutoPreTurn,
-            pressure_level: alan_agent_protocol::CompactionPressureLevel::Soft,
-            result: MemoryFlushResult::Success,
-            skip_reason: None,
-            source_messages: Some(8),
-            output_path: Some(".alan/memory/daily/2026-03-03.md".to_string()),
-            warning_message: None,
-            error_message: None,
-            timestamp: "2026-03-03T10:05:00Z".to_string(),
-        };
-        let items = [
-            RolloutItem::MemoryFlushAttempt(first),
-            RolloutItem::MemoryFlushAttempt(second.clone()),
-        ];
-
-        assert_eq!(
-            latest_memory_flush_attempt_from_rollout_items(&items),
-            Some(second)
-        );
-    }
-
-    #[tokio::test]
-    async fn test_persist_compaction_attempt_updates_latest_and_rollout() {
-        let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-        let mut machine = AgentMachine::new_with_recorder_in_dir(
-            "/proc/test",
-            "gemini-2.0-flash",
-            temp_dir.path(),
-        )
-        .await
-        .unwrap();
-        let attempt = CompactionAttemptSnapshot {
-            attempt_id: "attempt-123".to_string(),
-            submission_id: Some("sub-456".to_string()),
-            request: CompactionRequestMetadata {
-                mode: CompactionMode::Manual,
-                trigger: CompactionTrigger::Manual,
-                reason: CompactionReason::ExplicitRequest,
-                focus: Some("preserve todos".to_string()),
-            },
-            result: CompactionResult::Success,
-            pressure_level: None,
-            memory_flush_attempt_id: None,
-            input_messages: Some(10),
-            output_messages: Some(3),
-            input_prompt_tokens: Some(800),
-            output_prompt_tokens: Some(250),
-            retry_count: 0,
-            tape_mutated: true,
-            warning_message: None,
-            error_message: None,
-            failure_streak: None,
-            reference_context_revision_before: Some(2),
-            reference_context_revision_after: Some(2),
-            timestamp: "2026-03-03T10:00:00Z".to_string(),
-        };
-
-        machine
-            .persist_compaction_observation(attempt.clone(), None)
-            .await
-            .unwrap();
-        assert_eq!(machine.latest_compaction_attempt(), Some(&attempt));
-
-        let rollout_path = machine.rollout_path().unwrap().clone();
-        let items = RolloutRecorder::load_history(&rollout_path).await.unwrap();
-        let persisted = items.into_iter().find_map(|item| match item {
-            RolloutItem::CompactionAttempt(snapshot) => Some(snapshot),
-            _ => None,
-        });
-
-        assert_eq!(persisted, Some(attempt));
-    }
-
-    #[tokio::test]
-    async fn test_persist_memory_flush_attempt_updates_latest_and_rollout() {
-        let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-        let mut machine = AgentMachine::new_with_recorder_in_dir(
-            "/proc/test",
-            "gemini-2.0-flash",
-            temp_dir.path(),
-        )
-        .await
-        .unwrap();
-        let attempt = MemoryFlushAttemptSnapshot {
-            attempt_id: "flush-123".to_string(),
-            compaction_mode: CompactionMode::AutoPreTurn,
-            pressure_level: alan_agent_protocol::CompactionPressureLevel::Soft,
-            result: MemoryFlushResult::Success,
-            skip_reason: None,
-            source_messages: Some(7),
-            output_path: Some(".alan/memory/daily/2026-03-03.md".to_string()),
-            warning_message: None,
-            error_message: None,
-            timestamp: "2026-03-03T10:00:00Z".to_string(),
-        };
-
-        machine
-            .persist_memory_flush_attempt(attempt.clone())
-            .await
-            .unwrap();
-        assert_eq!(machine.latest_memory_flush_attempt(), Some(&attempt));
-
-        let rollout_path = machine.rollout_path().unwrap().clone();
-        let items = RolloutRecorder::load_history(&rollout_path).await.unwrap();
-        let persisted = items.into_iter().find_map(|item| match item {
-            RolloutItem::MemoryFlushAttempt(snapshot) => Some(snapshot),
-            _ => None,
-        });
-
-        assert_eq!(persisted, Some(attempt));
-    }
-
-    #[tokio::test]
-    async fn test_persist_compaction_observation_batches_attempt_and_summary() {
-        let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-        let mut machine = AgentMachine::new_with_recorder_in_dir(
-            "/proc/test",
-            "gemini-2.0-flash",
-            temp_dir.path(),
-        )
-        .await
-        .unwrap();
-        let attempt = CompactionAttemptSnapshot {
-            attempt_id: "attempt-batched".to_string(),
-            submission_id: Some("sub-batched".to_string()),
-            request: CompactionRequestMetadata {
-                mode: CompactionMode::Manual,
-                trigger: CompactionTrigger::Manual,
-                reason: CompactionReason::ExplicitRequest,
-                focus: Some("preserve blockers".to_string()),
-            },
-            result: CompactionResult::Retry,
-            pressure_level: None,
-            memory_flush_attempt_id: None,
-            input_messages: Some(10),
-            output_messages: Some(3),
-            input_prompt_tokens: Some(800),
-            output_prompt_tokens: Some(250),
-            retry_count: 1,
-            tape_mutated: true,
-            warning_message: None,
-            error_message: None,
-            failure_streak: None,
-            reference_context_revision_before: Some(2),
-            reference_context_revision_after: Some(2),
-            timestamp: "2026-03-03T10:00:00Z".to_string(),
-        };
-        let compacted = CompactedItem {
-            message: "Summary after retry".to_string(),
-            attempt_id: Some(attempt.attempt_id.clone()),
-            trigger: Some(CompactionTrigger::Manual),
-            reason: Some(CompactionReason::ExplicitRequest),
-            focus: Some("preserve blockers".to_string()),
-            input_messages: Some(10),
-            output_messages: Some(3),
-            input_tokens: Some(800),
-            output_tokens: Some(250),
-            duration_ms: Some(35),
-            retry_count: Some(1),
-            result: Some(CompactionResult::Retry),
-            reference_context_revision: Some(2),
-            timestamp: "2026-03-03T10:00:01Z".to_string(),
-        };
-
-        machine
-            .persist_compaction_observation(attempt.clone(), Some(compacted.clone()))
-            .await
-            .unwrap();
-
-        let rollout_path = machine.rollout_path().unwrap().clone();
-        let items = RolloutRecorder::load_history(&rollout_path).await.unwrap();
-        let persisted_attempt = items.iter().find_map(|item| match item {
-            RolloutItem::CompactionAttempt(snapshot) => Some(snapshot),
-            _ => None,
-        });
-        let persisted_compacted = items.iter().find_map(|item| match item {
-            RolloutItem::Compacted(compacted) => Some(compacted),
-            _ => None,
-        });
-
-        assert_eq!(machine.latest_compaction_attempt(), Some(&attempt));
-        assert_eq!(persisted_attempt, Some(&attempt));
-        assert_eq!(
-            persisted_compacted.map(|item| item.attempt_id.as_deref()),
-            Some(Some("attempt-batched"))
-        );
-        assert_eq!(
-            persisted_compacted.map(|item| item.message.as_str()),
-            Some(compacted.message.as_str())
-        );
-    }
-
-    #[test]
-    fn test_record_turn_context() {
-        let machine = AgentMachine::new();
-        let context_items = vec![ContextItem {
-            id: "ctx-1".to_string(),
-            kind: "customer".to_string(),
-            title: "Customer Profile".to_string(),
-            content: "Test content".to_string(),
-            fingerprint: "abc123".to_string(),
-        }];
-        // Should not panic without recorder
-        machine.record_turn_context(
-            "gemini-2.0-flash",
-            Some(alan_agent_protocol::ReasoningEffort::Low),
-            "System prompt",
-            &context_items,
-            &["tool1".to_string(), "tool2".to_string()],
-            true,
-            &["skill1".to_string()],
-        );
-    }
-
-    #[test]
-    fn test_record_turn_context_if_changed_dedupes_identical_snapshots() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-            let mut machine = AgentMachine::new_with_recorder_in_dir(
-                "/proc/test",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-
-            let context_items = vec![ContextItem {
-                id: "ctx-1".to_string(),
-                kind: "customer".to_string(),
-                title: "Customer Profile".to_string(),
-                content: "Test content".to_string(),
-                fingerprint: "abc123".to_string(),
-            }];
-
-            let unchanged = ContextItemsDelta::default();
-            assert!(machine.record_turn_context_if_changed(
-                "gemini-2.0-flash",
-                None,
-                "System prompt",
-                &context_items,
-                &["tool1".to_string()],
-                true,
-                &["skill1".to_string()],
-                &unchanged,
-            ));
-
-            assert!(!machine.record_turn_context_if_changed(
-                "gemini-2.0-flash",
-                None,
-                "System prompt",
-                &context_items,
-                &["tool1".to_string()],
-                true,
-                &["skill1".to_string()],
-                &unchanged,
-            ));
-
-            // A tool list change should still record even when reference context is unchanged.
-            assert!(machine.record_turn_context_if_changed(
-                "gemini-2.0-flash",
-                None,
-                "System prompt",
-                &context_items,
-                &["tool1".to_string(), "tool2".to_string()],
-                true,
-                &["skill1".to_string()],
-                &unchanged,
-            ));
-
-            machine.flush().await;
-
-            let rollout_path = machine.rollout_path().unwrap().clone();
-            let items = RolloutRecorder::load_history(&rollout_path).await.unwrap();
-            let turn_context_count = items
-                .into_iter()
-                .filter(|item| matches!(item, RolloutItem::TurnContext(_)))
-                .count();
-            assert_eq!(turn_context_count, 2);
-        });
-    }
-
-    #[test]
-    fn test_record_turn_context_if_changed_records_reasoning_effort_changes() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-            let mut machine = AgentMachine::new_with_recorder_in_dir(
-                "/proc/test",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-
-            let context_items = vec![ContextItem {
-                id: "ctx-1".to_string(),
-                kind: "customer".to_string(),
-                title: "Customer Profile".to_string(),
-                content: "Test content".to_string(),
-                fingerprint: "abc123".to_string(),
-            }];
-            let unchanged = ContextItemsDelta::default();
-
-            assert!(machine.record_turn_context_if_changed(
-                "gemini-2.0-flash",
-                Some(alan_agent_protocol::ReasoningEffort::Low),
-                "System prompt",
-                &context_items,
-                &["tool1".to_string()],
-                true,
-                &["skill1".to_string()],
-                &unchanged,
-            ));
-            assert!(machine.record_turn_context_if_changed(
-                "gemini-2.0-flash",
-                Some(alan_agent_protocol::ReasoningEffort::High),
-                "System prompt",
-                &context_items,
-                &["tool1".to_string()],
-                true,
-                &["skill1".to_string()],
-                &unchanged,
-            ));
-
-            machine.flush().await;
-
-            let rollout_path = machine.rollout_path().unwrap().clone();
-            let efforts = RolloutRecorder::load_history(&rollout_path)
-                .await
-                .unwrap()
-                .into_iter()
-                .filter_map(|item| match item {
-                    RolloutItem::TurnContext(ctx) => Some(ctx.reasoning_effort),
-                    _ => None,
-                })
-                .collect::<Vec<_>>();
-            assert_eq!(
-                efforts,
-                vec![
-                    Some(alan_agent_protocol::ReasoningEffort::Low),
-                    Some(alan_agent_protocol::ReasoningEffort::High)
-                ]
-            );
-        });
-    }
-
-    #[test]
-    fn test_add_tool_message_persists_redacted_tool_payload_with_tool_call_id() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            use tokio::time::{Duration, Instant, sleep};
-
-            let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-            let mut machine = AgentMachine::new_with_recorder_in_dir(
-                "/proc/test",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-
-            machine.add_tool_message(
-                "call_789",
-                "web_search",
-                serde_json::json!({
-                    "ok": true,
-                    "headers": {
-                        "set-cookie": "machine=secret-cookie",
-                        "content-type": "application/json"
-                    }
-                }),
-            );
-
-            let rollout_path = machine.rollout_path().unwrap().clone();
-            let start = Instant::now();
-            let mut found = false;
-            while start.elapsed() < Duration::from_secs(1) {
-                if let Ok(content) = tokio::fs::read_to_string(&rollout_path).await
-                    && content.contains("\"role\":\"tool\"")
-                    && content.contains("\"tool_name\":\"call_789\"")
-                    && content.contains("\\\"set-cookie\\\":\\\"[REDACTED reason=secret_key]\\\"")
-                    && !content.contains("secret-cookie")
-                {
-                    found = true;
-                    break;
-                }
-                sleep(Duration::from_millis(10)).await;
-            }
-            assert!(
-                found,
-                "expected tool message with payload and tool_call_id to be persisted"
-            );
-        });
-    }
-
-    #[tokio::test]
-    async fn test_flush_waits_for_queued_rollout_writes() {
-        let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-
-        let mut machine = AgentMachine::new_with_recorder_in_dir(
-            "/proc/test",
-            "gemini-2.0-flash",
-            temp_dir.path(),
-        )
-        .await
-        .unwrap();
-        machine.add_user_message("u1");
-        machine.add_assistant_message("a1", None);
-        machine.record_event("evt", serde_json::json!({"ok": true}));
-        machine.flush().await;
-
-        let rollout_path = machine.rollout_path().unwrap().clone();
-        let content = tokio::fs::read_to_string(&rollout_path).await.unwrap();
-        let user_pos = content.find("\"content\":\"u1\"").unwrap();
-        let assistant_pos = content.find("\"content\":\"a1\"").unwrap();
-        let event_pos = content.find("\"event_type\":\"evt\"").unwrap();
-        assert!(user_pos < assistant_pos);
-        assert!(assistant_pos < event_pos);
-    }
-
-    #[tokio::test]
-    async fn test_rollback_records_non_durable_audit_marker() {
-        let temp_dir = TempDir::new_in(std::env::temp_dir()).unwrap();
-
-        let mut machine = AgentMachine::new_with_recorder_in_dir(
-            "/proc/test",
-            "gemini-2.0-flash",
-            temp_dir.path(),
-        )
-        .await
-        .unwrap();
-        machine.add_user_message("u1");
-        machine.add_assistant_message("a1", None);
-        machine.add_user_message("u2");
-        machine.add_assistant_message("a2", None);
-
-        let removed = machine.rollback_last_turns(1);
-        assert_eq!(removed.removed_turns, 1);
-        assert_eq!(removed.removed_messages, 2);
-        machine.flush().await;
-
-        let rollout_path = machine.rollout_path().unwrap().clone();
-        let items = RolloutRecorder::load_history(&rollout_path).await.unwrap();
-        let rollback_event = items.into_iter().find_map(|item| match item {
-            RolloutItem::Event(event) if event.event_type == "machine_rollback" => Some(event),
-            _ => None,
-        });
-
-        let event = rollback_event.expect("expected machine_rollback event");
-        assert_eq!(event.payload["requested_turns"], serde_json::json!(1));
-        assert_eq!(event.payload["removed_turns"], serde_json::json!(1));
-        assert_eq!(event.payload["removed_messages"], serde_json::json!(2));
-        assert_eq!(event.payload["durable"], serde_json::json!(false));
-        assert_eq!(event.payload["scope"], serde_json::json!("in_memory"));
-        assert_eq!(
-            event.payload["warning"],
-            serde_json::json!(ROLLBACK_NON_DURABLE_WARNING)
-        );
-    }
-
-    // Tests for payload truncation
-
-    #[test]
-    fn test_truncate_payload_small_payload_unchanged() {
-        let payload = serde_json::json!({
-            "success": true,
-            "url": "https://example.com",
-            "title": "Example"
-        });
-        let result = truncate_payload(payload.clone(), 1000);
-        assert_eq!(result, payload);
-    }
-
-    #[test]
-    fn test_truncate_text() {
-        let text = "This is a long text that needs to be truncated";
-        let truncated = truncate_text(text, 20);
-        assert!(truncated.contains("...[truncated]"));
-        assert!(truncated.len() < text.len() + 15); // +15 for "...[truncated]"
-    }
-
-    #[test]
-    fn test_truncate_text_short() {
-        let text = "Short";
-        let truncated = truncate_text(text, 100);
-        assert_eq!(truncated, text);
-    }
-
-    #[test]
-    fn test_truncate_payload_large_content() {
-        let large_content = "a".repeat(10000);
-        let payload = serde_json::json!({
-            "success": true,
-            "url": "https://example.com",
-            "content": large_content
-        });
-        let result = truncate_payload(payload, 5000);
-        let result_str = result.to_string();
-        assert!(result_str.len() < 6000); // Should be significantly reduced
-        assert!(result_str.contains("...[truncated]"));
-    }
-
-    #[test]
-    fn test_truncate_payload_preserves_critical_fields() {
-        let large_content = "x".repeat(5000);
-        let payload = serde_json::json!({
-            "success": false,
-            "error": "Some error",
-            "url": "https://example.com",
-            "title": "Test Title",
-            "content": large_content
-        });
-        let result = truncate_payload(payload, 2000);
-        // Critical fields should be preserved
-        assert_eq!(result["success"], false);
-        assert_eq!(result["error"], "Some error");
-        assert_eq!(result["url"], "https://example.com");
-        assert_eq!(result["title"], "Test Title");
-    }
-
-    #[test]
-    fn test_add_tool_message_preserves_large_payload_on_tape() {
-        let mut machine = AgentMachine::new();
-        let large_content = "x".repeat(50000);
-        let payload = serde_json::json!({
-            "success": true,
-            "content": large_content
-        });
-
-        machine.add_tool_message("call_456", "test_tool", payload);
-
-        let messages = machine.tape.messages();
-        assert_eq!(messages.len(), 1);
-        let responses = messages[0].tool_responses();
-        assert_eq!(responses.len(), 1);
-
-        // Tape should keep the full payload; projection handles truncation.
-        let response_str = serde_json::to_string(&responses[0].content).unwrap();
-        assert!(
-            response_str.len() > 50000,
-            "Payload should stay full on tape, got {} chars",
-            response_str.len()
-        );
-    }
-
-    // Additional truncation tests for better coverage
-
-    #[test]
-    fn test_truncate_payload_array() {
-        let payload = serde_json::json!([
-            {"id": 1, "content": "First item"},
-            {"id": 2, "content": "Second item"},
-            {"id": 3, "content": "Third item"}
-        ]);
-        // Small max_size to trigger truncation
-        let result = truncate_payload(payload, 100);
-        // Result should be an array
-        assert!(result.is_array());
-        let arr = result.as_array().unwrap();
-        // Should contain items but may have truncation note
-        assert!(!arr.is_empty());
-    }
-
-    #[test]
-    fn test_truncate_payload_nested_object() {
-        let payload = serde_json::json!({
-            "level1": {
-                "level2": {
-                    "data": "x".repeat(5000)
-                }
-            }
-        });
-        let result = truncate_payload(payload, 1000);
-        // Should preserve structure but truncate content
-        assert!(result.get("level1").is_some());
-    }
-
-    #[test]
-    fn test_truncate_payload_string_only() {
-        let payload = serde_json::Value::String("a".repeat(5000));
-        let result = truncate_payload(payload, 1000);
-        // String should be truncated
-        let result_str = result.as_str().unwrap();
-        assert!(result_str.len() < 5000);
-        assert!(result_str.contains("...[truncated]"));
-    }
-
-    #[test]
-    fn test_truncate_payload_aggregated_content() {
-        let large_content = "b".repeat(5000);
-        let payload = serde_json::json!({
-            "success": true,
-            "aggregated_content": large_content
-        });
-        let result = truncate_payload(payload, 2000);
-        // aggregated_content should be truncated
-        let content = result["aggregated_content"].as_str().unwrap();
-        assert!(content.len() < 5000);
-        assert!(content.contains("...[truncated]"));
-    }
-
-    #[test]
-    fn test_truncate_payload_array_truncation_note() {
-        // Create a large array that will trigger truncation
-        let items: Vec<serde_json::Value> = (0..100)
-            .map(|i| serde_json::json!({"id": i, "data": "x".repeat(100)}))
-            .collect();
-        let payload = serde_json::Value::Array(items);
-        let result = truncate_payload(payload, 500);
-        // Should contain truncation note in one of the items
-        let arr = result.as_array().unwrap();
-        let has_note = arr.iter().any(|item| {
-            item.get("_note")
-                .and_then(|n| n.as_str())
-                .map(|s| s.contains("omitted"))
-                .unwrap_or(false)
-        });
-        assert!(has_note, "Should have truncation note in array items");
-    }
-
-    #[test]
-    fn test_truncate_payload_object_truncated_field() {
-        // Create an object with many large fields
-        let mut map = serde_json::Map::new();
-        for i in 0..50 {
-            map.insert(
-                format!("field{}", i),
-                serde_json::Value::String("y".repeat(200)),
-            );
-        }
-        let payload = serde_json::Value::Object(map);
-        let result = truncate_payload(payload, 1000);
-        // Should have _truncated field
-        assert!(
-            result.get("_truncated").is_some(),
-            "Should have _truncated field for omitted fields"
-        );
-    }
-
-    #[test]
-    fn test_truncate_payload_mixed_types() {
-        let payload = serde_json::json!({
-            "string": "test",
-            "number": 42,
-            "bool": true,
-            "null": null,
-            "array": [1, 2, 3],
-            "nested": {"key": "value"}
-        });
-        let result = truncate_payload(payload, 1000);
-        // All types should be preserved
-        assert_eq!(result["string"], "test");
-        assert_eq!(result["number"], 42);
-        assert_eq!(result["bool"], true);
-        assert!(result["null"].is_null());
-        assert!(result["array"].is_array());
-        assert!(result["nested"].is_object());
-    }
-
-    #[test]
-    fn test_rollback_last_turns_removes_latest_turn_messages() {
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("u1");
-        machine.add_assistant_message("a1", None);
-        machine.add_tool_message("call1", "web_search", serde_json::json!({"ok": true}));
-        machine.add_user_message("u2");
-        machine.add_assistant_message("a2", None);
-
-        let removed = machine.rollback_last_turns(1);
-
-        assert_eq!(removed.removed_turns, 1);
-        assert_eq!(removed.removed_messages, 2);
-        let messages = machine.tape.messages();
-        assert_eq!(messages.len(), 3);
-        assert_eq!(messages[0].text_content(), "u1");
-        assert_eq!(messages[1].text_content(), "a1");
-        assert!(messages[2].is_tool());
-    }
-
-    #[test]
-    fn test_rollback_last_turns_clears_all_when_request_exceeds_history() {
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("u1");
-        machine.add_assistant_message("a1", None);
-        machine.has_active_task = true;
-
-        let removed = machine.rollback_last_turns(10);
-
-        assert_eq!(removed.removed_turns, 1);
-        assert_eq!(removed.removed_messages, 2);
-        assert!(machine.tape.messages().is_empty());
-        assert!(!machine.has_active_task);
-    }
-
-    #[test]
-    fn test_rollback_last_turns_ignores_control_user_messages_for_turn_boundaries() {
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("u1");
-        machine.add_assistant_message("a1", None);
-        machine.add_user_control_message_parts(vec![ContentPart::structured(serde_json::json!({
-            "checkpoint_id": "tool_escalation_call-1",
-            "checkpoint_type": "tool_escalation",
-            "choice": "approve",
-            "__alan_internal_control": {
-                "kind": "tool_escalation_confirmation",
-                "version": 1,
-                "source": "runtime/submission_handlers"
-            }
-        }))]);
-        machine.add_assistant_message("a2", None);
-
-        let removed = machine.rollback_last_turns(1);
-
-        assert_eq!(
-            removed.removed_messages, 4,
-            "rollback should anchor on the real user turn, not synthetic control messages"
-        );
-        assert_eq!(removed.removed_turns, 1);
-        assert!(machine.tape.messages().is_empty());
-    }
-
-    #[test]
-    fn test_rollback_last_turns_ignores_effect_replay_control_messages_for_turn_boundaries() {
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("u1");
-        machine.add_assistant_message("a1", None);
-        machine.add_user_control_message_parts(vec![ContentPart::structured(serde_json::json!({
-            "checkpoint_id": "effect_replay_call-1",
-            "checkpoint_type": "effect_replay_confirmation",
-            "choice": "approve",
-            "__alan_internal_control": {
-                "kind": "effect_replay_confirmation",
-                "version": 1,
-                "source": "runtime/submission_handlers"
-            }
-        }))]);
-        machine.add_assistant_message("a2", None);
-
-        let removed = machine.rollback_last_turns(1);
-
-        assert_eq!(
-            removed.removed_messages, 4,
-            "rollback should ignore effect replay control messages the same way as policy controls"
-        );
-        assert_eq!(removed.removed_turns, 1);
-        assert!(machine.tape.messages().is_empty());
-    }
-}
+#[path = "agent_machine_tests.rs"]
+mod tests;
